requirements.txt
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
black==22.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.2
colorama==0.4.4
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.3
django-allauth==0.50.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-extensions==3.2.0
django-filter==21.1
django-heroku==0.3.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.0
drf-spectacular-sidecar==2022.4.1
et-xmlfile==1.1.0
Faker==13.3.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
jsonschema==4.4.0
Markdown==3.3.6
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
openpyxl==3.0.9
pandas==1.4.1
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
toposort==1.7
typing_extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
whitenoise==6.0.0
zipp==3.8.0
